[edit] UserInfoItem 컴포넌트로 사용자 정보 로직 분리
